check 만들기

+++ src/main/java/com/example/kiwi/repository/SelectionRep.java
@@ -9,5 +9,6 @@ import java.util.Optional;
 
 @Repository
 @Transactional
-public interface SelectionRep extends JpaRepository<Selection, Integer> {
+public interface SelectionRep extends JpaRepository<Selection, Long> {
+
 }
